refactor(review-tooling): review-preflight and ci-triage move into skills/pr-review
Owner, 2026-09-04: "both review-preflight and ci-triage don't belong to
scripts/, imo." Moved as one skill (SKILL.md + scripts/), invoked by path;
neither is a boot dependency of the core.

- review-preflight's git-less fallback for the repo root is now three
  levels above its file (was two); both fallback tests updated.
- Callers repointed: AGENTS.md, CLAUDE.md, REVIEW.md's run line, the
  workspace-resolution lint's allowlist and comment, three test files.
- scripts/review-preflight.py and scripts/ci-triage.py remain as exec
  shims for one release, so the pr-triage skill (sutando-skills) and
  peers' notes keep working until repointed; both shims exercised.
- Not moved: scripts/review-checks.sh (CI runs it by path; a separate
  decision).

Suites: review-preflight (42), ci-triage-subjects, ci-triage-superseded-runs
all green; lint-workspace-resolution rc=0; review-checks clean on the diff.

Stand: Echo Act IV Mini
